agent: Fix signing data longer than ASSUAN_LINELENGTH when using cards
* call-scd.c (agent_card_pksign): splits indata when longer than
ASSUAN_LINELENGTH
When using SSH with certificates and card like yubikey, SSH asks the
card to sign the whole certificate.
The result data lenght to send to the card exceeds the buffer capacity

char line[ASSUAN_LINELENGTH];

defined in the function agent_card_pksign.
This produces a failure with a generic error.
This fix takes the code to split the line in smaller chunk from the
function agent_card_pkdecrypt.