Stronger deprecate the --supervised option.
* agent/gpg-agent.c (opts): Rename option supervised.
* dirmngr/dirmngr.c (opts): Ditto.
--

The --supervised way to start gpg-agent has been deprecated for 2.5
years and will probably entirely removed with version 2.6.0.  To allow
its use until its removal the systemd service description need to be
adjusted to use this option.  The reason for the deprecation are
conflicts with the way systemd starts gpg-agent and gpg expects how
gpg-agent is started.  In particular gpg expects that the gpg-agent
matching its own version is started.  Further the systemd way is not
portable to other platforms and long term experience on Windows has
show that the standard way of starting gpg-agent is less error prone.

Note to those who want to re-introduse this option: Pretty please do
not use socket names conflicting with our standard socket names.  For
example use /run/user/1000/foo-gnupg/S.gpg-agent.