agent: Fix ssh-agent's request_identities for skipped keys.
* agent/command-ssh.c (ssh_send_available_keys): Adjust key counter
for skipped keys.
--

Fixes-commit: 8b8a8b246c443d5631a88ec59b88edf00aa0ff51
which introduced a regression due to an extra variable for counting
the keys.

The bug showed up for example if a card with a Brainpool Auth key was
also used.  Unfortunately OpenSSH still does not allow for Brainpool
keys.

(cherry picked from commit 2469dc5aae671247100339493ad7919919a02db9)