scd,w32: Fix posssible lockup due to lost select results.
* scd/scdaemon.c (handle_connections) [W32]: Do not continue the loop
when an event was encountered.
--

Here the event handle is passed to npth_eselect so that this function
can detect the event and reset the event.  There is no need to consume
this information here.  However, npth_select might also got a ready
file descriptor along with the event and by doing a "continue" we
would miss the ready state of the file descriptor.  The fix is to do
nothing here, similar to what we do in gpg-agent.

Fixes-commit: f9acc7d18bb90f47dafe7e32ae92f567756d6b12
GnuPG-bug-id: 2982