gpgsm: Fix caching of the trustlist's flags.
* sm/call-agent.c (istrusted_status_cb): Clear the helper flags.
--

GnuPG-bug-id: 7738
Fixes-commit: 09d4b8f496dd461a21d5ba0297710d683b16def4
which introduced the caching but go the flags wrong.  The fallout from
this was that the qualified signature consent was show for non QES
certificates.  Other flags are almost always the same and thus no
other harm was done.

Backported-from-master: 67a3020d0f3beed89270f2d004c3da5ee9cbfc63