gpgtar: Fix regression in end-of-archive detection.
* tools/gpgtar-extract.c (gpgtar_extract): Detect end-of-archive and
break loop.
* tools/gpgtar-list.c (gpgtar_list): Ditto.  Also do not print
"running failed" if the exitcode is 0.
--

GnuPG-bug-id: 7757

In the old code (gnupg 2.4) we did not call gnupg_release_process when
cleaning up the gpg process and thus it could keep on running and
deliver its output to the status-fd.  The new code does a
gnupg_release_process which sends a SIGTERM to a not yet terminated
process and thus the output was cut off.