scd:openpgp: Support Yubikey attestion generation.
* scd/app-openpgp.c (data_objects): Add more v3 DOs.
(do_setattr): New name GEN-ATTST as pseudo attribute to generate a
Yubikey attestation certificate.
--

Alhough this command seems to work, I am only able to get the
OPENPGP.3 cert back.  This might be related to the bug described in
the commit 37b1c5c2004c1147a13b388863aaa8f0caf7d71f which implemented
the select data functionality and showed that it did not work on a
Yubikey.  This time the test target was a Yubikey 5.4.3.