agent: Enable "relax" in the trustlist by default and add "norelax"
* agent/trustlist.c (read_one_trustfile): Enabled the relax flag be
default.  Add new keyword "norelax".
--

In any real world settings the "relax" flag is always set.  Thus it is
better to make it the default and provide a way to disable it.  In
fact, it is questionable whether this flag make sense at all.  It is
some kind of legacy thing.